Move the worker and analyser from the daemon
This avoids having the code in a single file, making it more readable
and maintainable
