+++ src/tools/efimon-daemon/efimon-worker.cpp
@@ -1,0 +1,9 @@
+/**
+ * @file efimon-worker.cpp
+ * @author Luis G. Leon-Vega (luis.leon@ieee.org)
+ * @brief Defines the process daemon attachable worker
+ *
+ * @copyright Copyright (c) 2024. See License for Licensing
+ */
+
+#include "efimon-daemon/efimon-worker.hpp"  // NOLINT
